removd hosted option from empty blazor wasm template
