--- src/ProjectTemplates/Web.ProjectTemplates/EmptyComponentsWebAssembly-CSharp.Server.csproj.in
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>${DefaultNetCoreTargetFramework}</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <NoDefaultLaunchSettingsFile Condition="'$(ExcludeLaunchSettings)' == 'True'">True</NoDefaultLaunchSettingsFile>
-    <RootNamespace Condition="'$(name)' != '$(name{-VALUE-FORMS-}safe_namespace)'">EmptyComponentsWebAssembly-CSharp.Server</RootNamespace>
-    <AssemblyName Condition="'$(name)' != '$(name{-VALUE-FORMS-}safe_namespace)'">`$(AssemblyName.Replace(' ', '_'))</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="${MicrosoftAspNetCoreComponentsWebAssemblyServerVersion}" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Client\EmptyComponentsWebAssembly-CSharp.Client.csproj" />
-    <ProjectReference Include="..\Shared\EmptyComponentsWebAssembly-CSharp.Shared.csproj" />
-  </ItemGroup>
-
-</Project>
