--- src/ProjectTemplates/Web.ProjectTemplates/EmptyComponentsWebAssembly-CSharp.Shared.csproj.in
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>${DefaultNetCoreTargetFramework}</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <SupportedPlatform Include="browser" />
-  </ItemGroup>
-</Project>
