+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/dotnetcli.host.json
@@ -5,9 +5,6 @@
       "longName": "no-restore",
       "shortName": ""
     },
-    "Hosted": {
-      "longName": "hosted"
-    },
     "PWA": {
       "longName": "pwa"
     },
