+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.cs.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Cílová architektura pro projekt",
   "symbols/Framework/choices/net8.0/description": "Cílový net8.0",
   "symbols/skipRestore/description": "Pokud se tato možnost zadá, přeskočí automatické obnovení projektu při vytvoření.",
-  "symbols/Hosted/displayName": "_Hostováno pomocí ASP.NET Core",
-  "symbols/Hosted/description": "Pokud je tato možnost zadaná, zahrnuje hostitele ASP.NET Core pro aplikaci Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Určuje, jestli se má z vygenerované šablony vyloučit soubor launchSettings.json.",
   "symbols/kestrelHttpPort/description": "Číslo portu, který se má použít pro koncový bod HTTP v souboru launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Číslo portu, který se má použít pro koncový bod HTTPS v souboru launchSettings.json. Tato možnost se dá použít jenom v případě, že se nepoužívá parametr no-https.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Pokud je tato možnost zadaná, vytvoří progresivní webovou aplikaci (PWA), která podporuje instalaci a offline použití.",
   "symbols/HasHttpProfile/description": "Vždy mějte profil protokolu HTTP.",
   "symbols/NoHttps/description": "Určuje, jestli se má vypnout protokol HTTPS.",
-  "postActions/restore/description": "Obnoví balíčky NuGet vyžadované tímto projektem.",
-  "postActions/restore/manualInstructions/default/text": "Spustit dotnet restore",
   "postActions/restoreClient/description": "Obnoví balíčky NuGet vyžadované tímto projektem.",
   "postActions/restoreClient/manualInstructions/default/text": "Spustit „dotnet restore“"
 }
