+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.es.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Marco de destino del proyecto.",
   "symbols/Framework/choices/net8.0/description": "net8.0 de destino",
   "symbols/skipRestore/description": "Si se especifica, se omite la restauración automática del proyecto durante la creación.",
-  "symbols/Hosted/displayName": "ASP.NET Core _hospedado",
-  "symbols/Hosted/description": "Si se especifica, incluye un host de ASP.NET Core para la aplicación Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Indica si se va a excluir launchSettings.json de la plantilla generada.",
   "symbols/kestrelHttpPort/description": "Número de puerto que se va a usar para el punto de conexión HTTP en launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Número de puerto que se va a usar para el punto de conexión HTTPS en launchSettings.json. Esta opción solo es aplicable cuando no se usa el parámetro no-https.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Si se especifica, produce una aplicación web progresiva (PWA) compatible con la instalación y el uso sin conexión.",
   "symbols/HasHttpProfile/description": "Siempre tiene un perfil HTTP.",
   "symbols/NoHttps/description": "Si se va a desactivar HTTPS.",
-  "postActions/restore/description": "Restaure los paquetes NuGet necesarios para este proyecto.",
-  "postActions/restore/manualInstructions/default/text": "Ejecutar \"dotnet restore\"",
   "postActions/restoreClient/description": "Restaure los paquetes NuGet necesarios para este proyecto.",
   "postActions/restoreClient/manualInstructions/default/text": "Ejecutar \"dotnet restore\""
 }
