+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.fr.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Framework cible du projet.",
   "symbols/Framework/choices/net8.0/description": "Cible net8.0",
   "symbols/skipRestore/description": "S’il est spécifié, ignore la restauration automatique du projet lors de la création.",
-  "symbols/Hosted/displayName": "ASP.NET core _Hosted",
-  "symbols/Hosted/description": "S’il est spécifié, inclut un hôte ASP.NET Core pour l’application Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Indique s’il faut exclure launchSettings.json du modèle généré.",
   "symbols/kestrelHttpPort/description": "Numéro de port à utiliser pour le point de terminaison HTTP dans launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Numéro de port à utiliser pour le point de terminaison HTTPS dans launchSettings.json. Cette option s’applique uniquement lorsque le paramètre no-https n’est pas utilisé.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Si ce paramètre est spécifié, produit une application web progressive (PWA) qui prend en charge l’installation et l’utilisation hors connexion.",
   "symbols/HasHttpProfile/description": "Toujours avoir un profil HTTP.",
   "symbols/NoHttps/description": "Indique s’il faut désactiver HTTPS.",
-  "postActions/restore/description": "Restaurez les packages NuGet requis par ce projet.",
-  "postActions/restore/manualInstructions/default/text": "Exécuter « dotnet restore »",
   "postActions/restoreClient/description": "Restaurez les packages NuGet requis par ce projet.",
   "postActions/restoreClient/manualInstructions/default/text": "Exécuter « dotnet restore »"
 }
