+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.it.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Il framework di destinazione per il progetto.",
   "symbols/Framework/choices/net8.0/description": "Destinazione net8.0",
   "symbols/skipRestore/description": "Se specificato, ignora il ripristino automatico del progetto durante la creazione.",
-  "symbols/Hosted/displayName": "ASP.NET Core _Hosted",
-  "symbols/Hosted/description": "Se specificato, include un host ASP.NET Core per l'app Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Indica se escludere launchSettings.json dal modello generato.",
   "symbols/kestrelHttpPort/description": "Numero di porta da usare per l'endpoint HTTP in launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Numero di porta da usare per l'endpoint HTTPS in launchSettings.json. Questa opzione è applicabile solo quando il parametro no-https non viene usato.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Se specificato, produce un'applicazione Web progressiva (PWA) che supporta l'installazione e l'uso offline.",
   "symbols/HasHttpProfile/description": "Avere sempre un profilo HTTP.",
   "symbols/NoHttps/description": "Indica se disattivare HTTPS.",
-  "postActions/restore/description": "Ripristina i pacchetti NuGet richiesti da questo progetto.",
-  "postActions/restore/manualInstructions/default/text": "Esegui 'dotnet restore'",
   "postActions/restoreClient/description": "Ripristina i pacchetti NuGet richiesti da questo progetto.",
   "postActions/restoreClient/manualInstructions/default/text": "Esegui 'dotnet restore'"
 }
