+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.ko.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "프로젝트에 대한 대상 프레임워크입니다.",
   "symbols/Framework/choices/net8.0/description": "대상 net8.0",
   "symbols/skipRestore/description": "지정된 경우, 프로젝트 생성 시 자동 복원을 건너뜁니다.",
-  "symbols/Hosted/displayName": "ASP.NET Core 호스팅(_H)",
-  "symbols/Hosted/description": "지정된 경우 Blazor WebAssembly 앱에 대한 ASP.NET Core 호스트를 포함합니다.",
   "symbols/ExcludeLaunchSettings/description": "생성된 템플릿에서 launchSettings.json을 제외할지 여부입니다.",
   "symbols/kestrelHttpPort/description": "launchSettings.json의 HTTP 엔드포인트에 사용할 포트 번호입니다.",
   "symbols/kestrelHttpsPort/description": "launchSettings.json의 HTTPS 엔드포인트에 사용할 포트 번호입니다. 이 옵션은 매개 변수 no-https가 사용되지 않는 경우에만 적용됩니다.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "지정된 경우 설치 및 오프라인 사용을 지원하는 PWA(프로그레시브 웹 응용 프로그램)를 생성합니다.",
   "symbols/HasHttpProfile/description": "항상 HTTP 프로필이 있어야 합니다.",
   "symbols/NoHttps/description": "HTTPS를 끌지 여부입니다.",
-  "postActions/restore/description": "이 프로젝트에 필요한 NuGet 패키지를 복원합니다.",
-  "postActions/restore/manualInstructions/default/text": "'dotnet restore' 실행",
   "postActions/restoreClient/description": "이 프로젝트에 필요한 NuGet 패키지를 복원합니다.",
   "postActions/restoreClient/manualInstructions/default/text": "'dotnet restore' 실행"
 }
