+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.pl.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Platforma docelowa dla tego projektu.",
   "symbols/Framework/choices/net8.0/description": "Docelowa platforma net8.0",
   "symbols/skipRestore/description": "Jeśli ta opcja jest określona, pomija automatyczne przywracanie projektu podczas tworzenia.",
-  "symbols/Hosted/displayName": "ASP.NET Core _Hosted",
-  "symbols/Hosted/description": "Jeśli zostanie określony, zawiera hosta platformy ASP.NET Core dla aplikacji Zestaw WebAssembly platformy Blazor.",
   "symbols/ExcludeLaunchSettings/description": "Określa, czy wykluczyć plik launchSettings.json z wygenerowanego szablonu.",
   "symbols/kestrelHttpPort/description": "Numer portu do użycia dla punktu końcowego HTTP w pliku launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Numer portu do użycia dla punktu końcowego HTTPS w pliku launchSettings.json. Ta opcja ma zastosowanie tylko wtedy, gdy parametr no-https nie jest używany.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Jeśli zostanie określony, tworzy progresywną aplikację internetową (PWA) obsługującą instalację i używanie w trybie offline.",
   "symbols/HasHttpProfile/description": "Zawsze masz profil HTTP.",
   "symbols/NoHttps/description": "Określa, czy wyłączyć protokół HTTPS.",
-  "postActions/restore/description": "Przywróć pakiety NuGet wymagane przez ten projekt.",
-  "postActions/restore/manualInstructions/default/text": "Uruchom polecenie \"dotnet restore\"",
   "postActions/restoreClient/description": "Przywróć pakiety NuGet wymagane przez ten projekt.",
   "postActions/restoreClient/manualInstructions/default/text": "Uruchom polecenie „dotnet restore”"
 }
