+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.pt-BR.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "A estrutura de destino do projeto.",
   "symbols/Framework/choices/net8.0/description": "net8.0 de destino",
   "symbols/skipRestore/description": "Se especificado, ignora a restauração automática do projeto sendo criado.",
-  "symbols/Hosted/displayName": "ASP.NET Core _Hosted",
-  "symbols/Hosted/description": "Se especificado, inclui um host ASP.NET Core para o aplicativo Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Se deve excluir launchSettings.json do modelo gerado.",
   "symbols/kestrelHttpPort/description": "Número da porta a ser usada para o ponto de extremidade HTTP em launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Número da porta a ser usada para o ponto de extremidade HTTPS em launchSettings.json. Esta opção só é aplicável quando o parâmetro no-https não é usado.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Se especificado, produz um Progressive Web Application (PWA) com suporte para instalação e uso offline.",
   "symbols/HasHttpProfile/description": "Sempre ter perfil HTTP.",
   "symbols/NoHttps/description": "Se o HTTPS deve ser desativado.",
-  "postActions/restore/description": "Restaure os pacotes NuGet exigidos por este projeto.",
-  "postActions/restore/manualInstructions/default/text": "Executar 'dotnet restore'",
   "postActions/restoreClient/description": "Restaure os pacotes NuGet exigidos por este projeto.",
   "postActions/restoreClient/manualInstructions/default/text": "Executa 'dotnet restore'"
 }
