+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.ru.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Целевая платформа для проекта.",
   "symbols/Framework/choices/net8.0/description": "Целевая net8.0",
   "symbols/skipRestore/description": "Если установлено, автоматическое восстановление проекта при создании пропускается.",
-  "symbols/Hosted/displayName": "_Размещение на ASP.NET Core",
-  "symbols/Hosted/description": "Если указывается, включает узел ASP.NET Core для приложения Blazor WebAssembly.",
   "symbols/ExcludeLaunchSettings/description": "Следует ли исключить launchSettings.json из созданного шаблона.",
   "symbols/kestrelHttpPort/description": "Номер порта, используемый для конечной точки HTTP в launchSettings.json.",
   "symbols/kestrelHttpsPort/description": "Номер порта, используемый для конечной точки HTTPS в launchSettings.json. Этот параметр применим только в том случае, если no-https не используется.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Если указывается, используется для создания прогрессивного веб-приложения (PWA), поддерживающего установку и автономное использование.",
   "symbols/HasHttpProfile/description": "Всегда имеет профиль HTTP.",
   "symbols/NoHttps/description": "Следует ли отключить HTTPS.",
-  "postActions/restore/description": "Восстановление пакетов NuGet, необходимых для этого проекта.",
-  "postActions/restore/manualInstructions/default/text": "Выполнить команду \"dotnet restore\"",
   "postActions/restoreClient/description": "Восстановление пакетов NuGet, необходимых для этого проекта.",
   "postActions/restoreClient/manualInstructions/default/text": "Выполнить команду \"dotnet restore\""
 }
