+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.tr.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "Projenin hedef çerçevesi.",
   "symbols/Framework/choices/net8.0/description": "Hedef net8.0",
   "symbols/skipRestore/description": "Belirtilirse, oluşturma sırasında projenin otomatik geri yüklenmesini atlar.",
-  "symbols/Hosted/displayName": "ASP.NET Core tarafından _Barındırılan",
-  "symbols/Hosted/description": "Belirtilmişse Blazor WebAssembly uygulaması için ASP.NET Core konağı içerir.",
   "symbols/ExcludeLaunchSettings/description": "launchSettings.json öğesinin oluşturulan şablondan dışlanıp dışlanmayacağı.",
   "symbols/kestrelHttpPort/description": "launchSettings.json içinde HTTP uç noktası için kullanılacak bağlantı noktası numarası.",
   "symbols/kestrelHttpsPort/description": "launchSettings.json içindeki HTTPS uç noktası için bağlantı noktası numarası. Bu seçenek yalnızca no-https parametresi kullanılmazsa uygulanabilir.",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "Belirtilmişse, yükleme ve çevrimdışı kullanımı destekleyen bir Aşamalı Web Uygulaması (PWA) oluşturur.",
   "symbols/HasHttpProfile/description": "Her zaman HTTP profili vardır.",
   "symbols/NoHttps/description": "HTTPS'nin kapatılıp kapatılmayacağı.",
-  "postActions/restore/description": "Bu projenin gerektirdiği NuGet paketlerini geri yükleyin.",
-  "postActions/restore/manualInstructions/default/text": "'dotnet restore' çalıştır",
   "postActions/restoreClient/description": "Bu projenin gerektirdiği NuGet paketlerini geri yükleyin.",
   "postActions/restoreClient/manualInstructions/default/text": "'dotnet restore' çalıştır"
 }
