+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.zh-Hans.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "项目的目标框架。",
   "symbols/Framework/choices/net8.0/description": "目标 net8.0",
   "symbols/skipRestore/description": "如果指定，则在创建时跳过项目的自动还原。",
-  "symbols/Hosted/displayName": "ASP.NET Core 托管(_H)",
-  "symbols/Hosted/description": "如果指定，则包括 Blazor WebAssembly 应用的 ASP.NET Core 主机。",
   "symbols/ExcludeLaunchSettings/description": "是否从生成的模板中排除 launchSettings.json。",
   "symbols/kestrelHttpPort/description": "要用于 launchSettings.json 中 HTTP 终结点的端口号。",
   "symbols/kestrelHttpsPort/description": "要用于 launchSettings.json 中 HTTPS 终结点的端口号。仅当不使用参数 no-https 时，此选项才适用。",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "如果指定，则生成支持安装和脱机使用的渐进式 Web 应用程序(PWA)。",
   "symbols/HasHttpProfile/description": "始终具有 HTTP 配置文件。",
   "symbols/NoHttps/description": "是否禁用 HTTPS。",
-  "postActions/restore/description": "还原此项目所需的 NuGet 包。",
-  "postActions/restore/manualInstructions/default/text": "运行 \"dotnet restore\"",
   "postActions/restoreClient/description": "还原此项目所需的 NuGet 包。",
   "postActions/restoreClient/manualInstructions/default/text": "运行 “dotnet restore”"
 }
