+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/localize/templatestrings.zh-Hant.json
@@ -5,8 +5,6 @@
   "symbols/Framework/description": "專案的目標 Framework。",
   "symbols/Framework/choices/net8.0/description": "目標 net8.0",
   "symbols/skipRestore/description": "若指定，會在建立時跳過專案的自動還原。",
-  "symbols/Hosted/displayName": "ASP.NET Core 託管(_H)",
-  "symbols/Hosted/description": "若指定，會包含 Blazor WebAssembly 應用程式的 ASP.NET Core 主機。",
   "symbols/ExcludeLaunchSettings/description": "是否要從產生的範本排除 launchSettings.json。",
   "symbols/kestrelHttpPort/description": "launchSettings.json 中 HTTP 端點要使用的連接埠號碼。",
   "symbols/kestrelHttpsPort/description": "launchSettings.json 中用於 HTTPS 端點的連接埠號碼。只有在未使用 no-https 參數時，才適用此選項。",
@@ -16,8 +14,6 @@
   "symbols/PWA/description": "若指定，會產生漸進式 Web 應用程式 (PWA) 支援安裝與離線使用。",
   "symbols/HasHttpProfile/description": "永遠有 HTTP 設定檔。",
   "symbols/NoHttps/description": "是否要關閉 HTTPS。",
-  "postActions/restore/description": "還原此專案所需的 NuGet 套件。",
-  "postActions/restore/manualInstructions/default/text": "執行 'dotnet restore'",
   "postActions/restoreClient/description": "還原此專案所需的 NuGet 套件。",
   "postActions/restoreClient/manualInstructions/default/text": "執行 'dotnet restore'"
 }
