+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/.template.config/template.json
@@ -18,24 +18,7 @@
   "preferNameDirectory": true,
   "primaryOutputs": [
     {
-      "condition": "(Hosted && (HostIdentifier == \"dotnetcli\" || HostIdentifier == \"dotnetcli-preview\"))",
-      "path": "EmptyComponentsWebAssembly-CSharp.sln"
-    },
-    {
-      "condition": "(Hosted && HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-      "path": "Server/EmptyComponentsWebAssembly-CSharp.Server.csproj"
-    },
-    {
-      "condition": "(!Hosted)",
       "path": "EmptyComponentsWebAssembly-CSharp.csproj"
-    },
-    {
-      "condition": "(Hosted && HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-      "path": "Client/EmptyComponentsWebAssembly-CSharp.Client.csproj"
-    },
-    {
-      "condition": "(Hosted && HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-      "path": "Shared/EmptyComponentsWebAssembly-CSharp.Shared.csproj"
     }
   ],
   "shortName": "blazorwasm-empty",
@@ -54,34 +37,15 @@
         {
           "condition": "(ExcludeLaunchSettings)",
           "exclude": [
-            "Client/Properties/launchSettings.json",
-            "Server/Properties/launchSettings.json"
-          ]
-        },
-        {
-          "condition": "(!Hosted)",
-          "exclude": [
-            "Server/**",
-            "Shared/**",
-            "*.sln"
-          ],
-          "rename": {
-            ".Client.csproj": ".csproj",
-            "Client": "."
-          }
-        },
-        {
-          "condition": "(Hosted && HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-          "exclude": [
-            "*.sln"
+            "Properties/launchSettings.json"
           ]
         },
         {
           "condition": "(!PWA)",
           "exclude": [
-            "Client/wwwroot/service-worker*.js",
-            "Client/wwwroot/manifest.webmanifest",
-            "Client/wwwroot/icon-512.png"
+            "wwwroot/service-worker*.js",
+            "wwwroot/manifest.webmanifest",
+            "wwwroot/icon-512.png"
           ]
         }
       ]
@@ -101,22 +65,11 @@
       "replaces": "net8.0",
       "defaultValue": "net8.0"
     },
-    "HostIdentifier": {
-      "type": "bind",
-      "binding": "HostIdentifier"
-    },
     "skipRestore": {
       "type": "parameter",
       "datatype": "bool",
       "description": "If specified, skips the automatic restore of the project on create.",
       "defaultValue": "false"
-    },
-    "Hosted": {
-      "type": "parameter",
-      "datatype": "bool",
-      "defaultValue": "false",
-      "displayName": "ASP.NET Core _Hosted",
-      "description": "If specified, includes an ASP.NET Core host for the Blazor WebAssembly app."
     },
     "ExcludeLaunchSettings": {
       "type": "parameter",
@@ -246,20 +199,8 @@
   },
   "postActions": [
     {
-      "id": "restore",
-      "condition": "(!skipRestore && Hosted)",
-      "description": "Restore NuGet packages required by this project.",
-      "manualInstructions": [
-        {
-          "text": "Run 'dotnet restore'"
-        }
-      ],
-      "actionId": "210D431B-A78B-4D2F-B762-4ED3E3EA9025",
-      "continueOnError": true
-    },
-    {
       "id": "restoreClient",
-      "condition": "(!skipRestore && !Hosted)",
+      "condition": "(!skipRestore)",
       "description": "Restore NuGet packages required by this project.",
       "manualInstructions": [
         {
@@ -267,7 +208,7 @@
         }
       ],
       "args": {
-        "files": [ "EmptyComponentsWebAssembly-CSharp.Client.csproj" ]
+        "files": [ "EmptyComponentsWebAssembly-CSharp.csproj" ]
       },
       "actionId": "210D431B-A78B-4D2F-B762-4ED3E3EA9025",
       "continueOnError": true
