--- src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/Shared/SharedClass.cs
@@ -1,1 +1,0 @@
-/* Shared classes can be referenced by both the Client and Server */
