+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/_Imports.razor
@@ -4,8 +4,4 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@*#if (!Hosted)
 @using EmptyComponentsWebAssembly_CSharp
-#else
-@using EmptyComponentsWebAssembly_CSharp.Client
-#endif*@
