+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/wwwroot/index.html
@@ -10,13 +10,8 @@
     <link href="manifest.webmanifest" rel="manifest" />
     <!--#endif -->
     
-    <!--#if Hosted -->
-    <!-- If you add any scoped CSS files, uncomment the following to load them
-    <link href="EmptyComponentsWebAssembly_CSharp.Client.styles.css" rel="stylesheet" /> -->
-    <!--#else -->
     <!-- If you add any scoped CSS files, uncomment the following to load them
     <link href="EmptyComponentsWebAssembly_CSharp.styles.css" rel="stylesheet" /> -->
-    <!--#endif -->
 </head>
 
 <body>
