+++ src/ProjectTemplates/Web.ProjectTemplates/content/EmptyComponentsWebAssembly-CSharp/wwwroot/manifest.webmanifest
