+++ src/ProjectTemplates/test/Templates.Tests/template-baselines.json
@@ -2230,25 +2230,6 @@
         "Program.cs"
       ]
     },
-    "Hosted": {
-      "Template": "blazorwasm-empty",
-      "Arguments": "new blazorwasm-empty --hosted",
-      "Files": [
-        "Client/_Imports.razor",
-        "Client/App.razor",
-        "Client/MainLayout.razor",
-        "Client/Program.cs",
-        "Client/Pages/Index.razor",
-        "Client/Properties/launchSettings.json",
-        "Client/wwwroot/index.html",
-        "Client/wwwroot/css/app.css",
-        "Server/Properties/launchSettings.json",
-        "Server/appsettings.Development.json",
-        "Server/appsettings.json",
-        "Server/Program.cs",
-        "Shared/SharedClass.cs"
-      ]
-    },
     "PWA": {
       "Template": "blazorwasm-empty",
       "Arguments": "new blazorwasm-empty --pwa",
@@ -2281,29 +2262,6 @@
         "Program.cs"
       ]
     },
-    "PWAHosted": {
-      "Template": "blazorwasm-empty",
-      "Arguments": "new blazorwasm-empty --pwa --hosted",
-      "Files": [
-        "Client/_Imports.razor",
-        "Client/App.razor",
-        "Client/MainLayout.razor",
-        "Client/Program.cs",
-        "Client/Pages/Index.razor",
-        "Client/Properties/launchSettings.json",
-        "Client/wwwroot/icon-512.png",
-        "Client/wwwroot/index.html",
-        "Client/wwwroot/manifest.webmanifest",
-        "Client/wwwroot/service-worker.js",
-        "Client/wwwroot/service-worker.published.js",
-        "Client/wwwroot/css/app.css",
-        "Server/Properties/launchSettings.json",
-        "Server/appsettings.Development.json",
-        "Server/appsettings.json",
-        "Server/Program.cs",
-        "Shared/SharedClass.cs"
-      ]
-    },
     "ExcludeLaunchSettings": {
       "Template": "blazorwasm-empty",
       "Arguments": "new blazorwasm-empty --exclude-launch-settings",
@@ -2315,23 +2273,6 @@
         "App.razor",
         "MainLayout.razor",
         "Program.cs"
-      ]
-    },
-    "ExcludeLaunchSettingsHosted": {
-      "Template": "blazorwasm-empty",
-      "Arguments": "new blazorwasm-empty --hosted --exclude-launch-settings",
-      "Files": [
-        "Client/_Imports.razor",
-        "Client/App.razor",
-        "Client/MainLayout.razor",
-        "Client/Program.cs",
-        "Client/Pages/Index.razor",
-        "Client/wwwroot/index.html",
-        "Client/wwwroot/css/app.css",
-        "Server/appsettings.Development.json",
-        "Server/appsettings.json",
-        "Server/Program.cs",
-        "Shared/SharedClass.cs"
       ]
     }
   }
